Test Building Nodes Extended + Minor Bugfixes
Test bulding extended with nodes and transitions (.Walls-file coming soon(tm)). More outside areas and floors were added, as well as a 2nd way to move between the floors.

Coordinates and IDs of nodes are now shown in the UI (mostly useful for debugging, and they can be disabled by setting a static variable to false).
Connections of transitions are now shown in the UI (useful for debugging, and they can be disabled by setting a static variable to false)
Node id is now also stored in the node-class.

Minor bugfixes:
- Duplicate transition removed
- Transition from node 0 to node 9 is now correctly connected from node 0 to node 1.
-  Pressing VK_M no longer reloads the simulation ("break;" was missing in the "switch")